update hash #412

+++ src/engine/board.hpp
@@ -73,7 +73,7 @@ class Board {
 #if USE_SIMD && USE_CRC32C_HASH
         // CRC32C Hash idea from http://www.amy.hi-ho.ne.jp/okuhara/edaxopt.htm#crc32hash
         inline uint32_t hash() const {
-            uint32_t res = _mm_crc32_u64(0, player);
+            uint64_t res = _mm_crc32_u64(0, player);
             res = _mm_crc32_u64(res, opponent);
             return global_hash_bit_mask & res;
         }
